Task 2: Extract agent/session_state.py with shared mutable globals
Moves SessionHandle, _active_sessions, _active_workers, _active_events,
_starting_workers, _global_session_semaphore, _shutdown_requested, callback
type aliases and dicts to a new thin module with no agent/ dependencies.

Re-exports all symbols from agent_session_queue for backward compat.
Updates write sites to use _session_state module attribute access for
primitive mutation correctness.
