+++ agent/agent_session_queue.py
@@ -25,8 +25,6 @@ import os
 import signal
 import subprocess
 import time
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
 from typing import Any
@@ -48,6 +46,24 @@ from agent.output_router import (
     determine_delivery_action,  # noqa: F401
 )
 from agent.session_logs import save_session_snapshot
+
+# Shared mutable session-tracking state — re-exported here for backward compatibility.
+import agent.session_state as _session_state  # noqa: F401 (also used for mutation sites)
+from agent.session_state import (  # noqa: F401
+    ReactionCallback,
+    ResponseCallback,
+    SendCallback,
+    SessionHandle,
+    _active_events,
+    _active_sessions,
+    _active_workers,
+    _global_session_semaphore,
+    _reaction_callbacks,
+    _response_callbacks,
+    _send_callbacks,
+    _shutdown_requested,
+    _starting_workers,
+)
 from agent.worktree_manager import WORKTREES_DIR, validate_workspace
 from config.enums import ClassificationType, PersonaType, SessionType
 from models.agent_session import AgentSession
@@ -2480,67 +2496,10 @@ def format_duration(seconds) -> str:
 # If no event fires within this window, the worker falls back to a sync query.
 DRAIN_TIMEOUT = 1.5  # seconds
 
-_active_workers: dict[str, asyncio.Task] = {}
-_active_events: dict[str, asyncio.Event] = {}
-
-
-@dataclass
-class SessionHandle:
-    """Per-session lookup handle for the two-tier no-progress detector (#1036).
-
-    Populated at the top of ``_execute_agent_session`` and removed in its
-    ``finally`` block. The ``task`` field is the ``asyncio.Task`` running the
-    session-scoped SDK work so the health check can cancel it when both Tier 1
-    heartbeats go stale AND all Tier 2 reprieve gates fail.
-
-    ``task`` is ``None`` until ``BackgroundTask.run()`` creates the
-    session-scoped task at ``agent/messenger.py:198``. Early setup code runs on
-    the worker-loop task itself; cancelling that would tear down the entire
-    worker (plan spike-1, #1039 review). Between registration and
-    ``BackgroundTask.run()`` there is nothing session-scoped to cancel — the
-    health check must check ``task is not None`` before calling ``.cancel()``.
-
-    The ``pid`` field is populated by the messenger's ``on_sdk_started``
-    callback once the SDK subprocess is spawned; it is used by the Tier 2
-    process-alive / has-children gates via psutil.
-
-    Lifecycle contract:
-      * Single writer: ``_execute_agent_session`` for its own session id.
-      * Multi reader: ``_agent_session_health_check`` and
-        ``_tier2_reprieve_signal`` (both look up by id, tolerate missing).
-      * ``_active_sessions.pop()`` happens in ``_execute_agent_session``'s
-        ``finally`` block, always.
-    """
-
-    task: asyncio.Task | None = None
-    pid: int | None = None
-
-
-# Registry of in-flight session executions, keyed by agent_session_id.
-#
-# Written only by ``_execute_agent_session`` (register at entry before any
-# raise site; pop in ``finally``). Read by the health check (to look up the
-# cancellable task on kill) and by ``_tier2_reprieve_signal`` (to read the
-# pid for Tier 2 gates).
-_active_sessions: dict[str, SessionHandle] = {}
-
-# Tracks worker_keys for which asyncio.create_task() has been called but the task
-# has not yet registered itself in _active_workers (i.e., the spawn is in-flight).
-# Since _ensure_worker() is synchronous (no await), check-and-set within the
-# function is atomic in the cooperative event loop, preventing duplicate workers
-# from being spawned when the health check iterates multiple pending sessions
-# sharing the same worker_key before either task is live in _active_workers.
-_starting_workers: set[str] = set()
-
-# Global concurrency ceiling: limits total simultaneously executing sessions
-# across all chat_ids. Initialized in _run_worker() before any worker loop
-# starts, so it is always available when _worker_loop() first awaits it.
-# None sentinel means no ceiling (pre-initialization or testing).
-_global_session_semaphore: asyncio.Semaphore | None = None
-
-# Graceful shutdown coordination: when set, worker loops finish their current
-# session and exit instead of waiting for new work.
-_shutdown_requested: bool = False
+# Mutable session-tracking globals live in agent.session_state (re-imported above).
+# _active_workers, _active_events, SessionHandle, _active_sessions, _starting_workers,
+# _global_session_semaphore, _shutdown_requested, _send_callbacks, _reaction_callbacks,
+# _response_callbacks are all imported from there.
 
 
 def request_shutdown() -> None:
@@ -2549,22 +2508,11 @@ def request_shutdown() -> None:
     Called by the standalone worker's SIGTERM handler. Sets the shutdown flag
     and wakes all waiting workers so they can check the flag and exit.
     """
-    global _shutdown_requested
-    _shutdown_requested = True
+    _session_state._shutdown_requested = True
     # Wake all waiting workers so they see the flag
     for event in _active_events.values():
         event.set()
     logger.info("Shutdown requested — workers will finish current sessions and exit")
-
-
-# Callbacks registered by the bridge for sending messages and reactions
-SendCallback = Callable[[str, str, int, Any], Awaitable[None]]  # (chat_id, text, reply_to, session)
-ReactionCallback = Callable[[str, int, str | None], Awaitable[None]]
-ResponseCallback = Callable[[object, str, str, int], Awaitable[None]]
-
-_send_callbacks: dict[str | tuple[str, str], SendCallback] = {}
-_reaction_callbacks: dict[str | tuple[str, str], ReactionCallback] = {}
-_response_callbacks: dict[str | tuple[str, str], ResponseCallback] = {}
 
 
 def register_callbacks(
@@ -2856,14 +2804,14 @@ async def _worker_loop(
     try:
         while True:
             # Check shutdown flag before starting new work
-            if _shutdown_requested:
+            if _session_state._shutdown_requested:
                 logger.info(f"[worker:{worker_key}] Shutdown requested, worker exiting")
                 break
 
             # Acquire global concurrency slot BEFORE popping — ensures that
             # transition_status("running") never occurs without a semaphore slot,
             # keeping the dashboard running count accurate.
-            semaphore = _global_session_semaphore
+            semaphore = _session_state._global_session_semaphore
             if semaphore is not None:
                 await semaphore.acquire()
             _semaphore_acquired = semaphore is not None
@@ -2905,7 +2853,7 @@ async def _worker_loop(
                 if standalone:
                     # Persistent mode: wait indefinitely for new work
                     await event.wait()
-                    if _shutdown_requested:
+                    if _session_state._shutdown_requested:
                         logger.info(f"[worker:{worker_key}] Woke from wait, shutdown requested")
                         break
                     # Re-acquire semaphore before retry pop
@@ -3179,7 +3127,7 @@ async def _worker_loop(
             event.clear()
 
             # Check shutdown flag after each completed session
-            if _shutdown_requested:
+            if _session_state._shutdown_requested:
                 logger.info(f"[worker:{worker_key}] Shutdown requested after session, exiting")
                 break
